+++ CHANGELOG.md
@@ -7,6 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added `tt` restart and status to the command palette.
+- Added `tt` install Tarantool CE to the command palette.
+
+### Changed
+
+- Now the package is called `tarantool` instead of `tarantool-vscode`.
+
 ## [0.0.1] - 28.03.2025
 
 ### Added